requirements.txt
Flask
Werkzeug
pyserial==3.4